+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file. This change
 ### Added
 - Information about contributors
 - Information about number of commits
+- Information about self-committed commits
 ### Changed
 - Include bootstrap for HTML output
 ### Removed